requirements.txt
# Agentic RAG — Python Dependencies
# Install with: pip install -r requirements.txt

# Environment variable management
python-dotenv>=1.0.0

# Core LangChain stack
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0

# LangGraph (agentic workflow)
langgraph>=0.1.0

# Vector store
chromadb>=0.5.0

# PDF loading
pypdf>=4.0.0

# Google Generative AI SDK
google-generativeai>=0.7.0

# OpenAI
langchain-openai>=0.1.0

# UI
streamlit>=1.35.0

# OCR & Image Processing (Phase 2)
# Requires system dependencies: Tesseract-OCR & Poppler
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0